luaG_errormsg should invoke msghandler to append stack trace to the error message even when no errfunc is specified. Also, since LUAI_THROW has suddenly stopped printing the error message in some cases, print error message before throwing.
Also, print error message before throwing Lua error: For some reason, luaD_throw -> LUAI_THROW is inconsistent about whether it prints the error message. When invoked through Inst::call, for some reason errorfunc is unset, and the error message is not printed.
As a hack, print the error message in luaG_errormsg before throwing.

Note: when lauG_errormsg is invoked as part of an error when running a file, errorfunc is set, but the stack trace is empty and the error message doesn't otherwise get printed.