Allow spell to be created with ValueMap args.
Convert castSpell, stopSpell to C wrapper, addCastWithArgs.
Move GObject::makeValueMapArg -> Lua::convert template.
Lua::inst: callIfExists with args (ended up not using, but tested it) and setGlobal<T> for assigning a C value and invoking the convert template.
cocos2d::Value: isNumber, isString.

note: one option was to pass valuemap args as parameter to spell init, but it's more convenient to set it as a global.