GSpace: nav bitmask for tracking obstacle tiles.
This is updated when shapes are created, meaning the methods for creating physics bodies shouldn't be static.
The space should have its size set before adding any bodies if the scene is going to use it; since the map scene doesn't open the map upon construction, it would be too ugly to require the GSpace to have its size set at construction.
Lua::convert: IntVec2 and std::vector