Interaction sensor
ObjectSensor mixin adapts the RadarObject mixin for use as a cone sensor to detect objects in front of the sensing agent. It has some hard-coded values as it is currently only used by the player.
objectSensor is a new physics object GType; currently it only senses enemies and enviroment objects. (An NPC class should be created).

InteractibleObject interface mixin
The Reimu NPC in the ClearingPath scene uses the interface to allow dialog.