InventoryObject
For now, all inventory objects are circle bodies, on the floor layer (both physics and graphics), radius 0.5, and ImageSprites.
InventoryObject wraps canInteract/interact interface with canAcquire/acquire interface for item classes.

GraveyardKey

Add floor GraphicsLayer.
Add floor PhysicsLayer.

Chipmunk++ fix: cp::Layers class tries to intepret the given layer number as an enumerated value and convert it to a bitmask value; it should accept the value as-is, since the GSpace API already expects the layer value is a bitmask.