Refactor controls
ControlRegister now tracks all control state. It is a singleton accessible through App.
ControlListener provides similar functionality to the defunct KeyListener. It simply provides a wrapper for adding callbacks, storing their UUIDs, and removing them when it is destroyed, i.e. when the ControlListener is destroyed along with the object that was using it, any callbacks referring to that object will be removed.
Instead of using Keys when checking control state or adding a callback, instead ControlActions are used. This is necessary since a keyboard key and gamepad button can each have more than one meaning, and these do not necessarily overlap.
Since a key or button can be mapped to multiple actions, ControlAction is actually a bitfield enumeration.
A ControlAction is considered to have a press if any corresponding button on the keyboard or gamepad is pressed, and is considered to have been released if none of those buttons is active - this involves comparing against the ControlAction state of the previous frame.
Refactored out code from Player for choosing between control stick state and movement or arrow key state. There is now a left and right vector. If a gamepad control stick is in a non-zero position (above deadzone threshold) it will be used for that vector, otherwise the corresponding keys will be used.

SpaceVect::isZero()
enum_bitwise_or
expand_enum_bitfield

Schedule App::update() - this allows ControlRegister to update regardless of GScene i.e. TitleMenu.