Anonymous GObjects and GObject identity rules
Anonymous GObjects do not need a name. It is not added to the GSpace name map. However, it is still used in the constructor, where the convention is to provide the class name. It is only available to be set when used with the “dynamic” GObject constructor
Includes Wall, Flower, and all Bullet objects.

GSpace will now suppress creation of invalid objects. An object is not considered valid if it has the same name of an existing object in the space (per the name map) and is marked not anonymous, or if it has a UUID that is in use. This involved refactoring processAdditions so that objects which will not be created can be skipped.

Anonymous objects will not have a Lua context created.
fixed GObject Lua API code to check anonymous flag and whether Lua ctx is initialized (in that order).