Refactor Lua/C API wrapper
wrap_cfunction: generates a function object (lambda) that calls a wrapped Lua-C function pointer
Lua::Inst::cfunctions now stores a map of function name to std::function<int(lua_State*)>
Lua::Inst::dispatch will be called for any wrapped function. It uses a C closure to store the name of the function. Thus, each cfunction object is a closure that has the ID of the functor