Spawner
Spawner is an Activateable object that will store and re-use its own args map as an argument to create object, i.e. objects wille created at the same position as the spawner. They will also have the same dimensions as the Spawner if it is a RectangleMapBody (not currently the case, as it is intended to be used for spawning enemies).

getSpawnArgs() - copies values stored at "spawn_type" "spawn_name" to "type", "name" so that Spawner args can be used to create object.