Menu layer controls, remove load/scene-select scenes
Move setup of control listener callbacks to MenuLayer.
MenuLayer::backPressed(), TextListMenuLayer::backPressed(), MapMenu::backPressed()
TextListMenuLayer can be popped with menuBack (ESC) if it is not the root menu (for menu scenes), and MapMenu can be closed with ESC.
Remove LoadProfileScene and SceneSelectScene - instead, these menus are pushed onto the menu stack in the TitleMenuScene.