Bullet/orb ricochet fix
Round collision normal to nearest 45° angle - part of the solution for the wall/barrier seam collision fix. This means it is still possible to get a diagonal normal when hitting an actual corner, as opposed to the seam between wall/barrier/door.

Fix colliding with inside of wall corner - Apply velocity change immediately instead of using message queue, but check that normal is in the correct direction - if object hits inside corner of walls, collisions with both wall pieces need to be processed in the same frame, and each one will transform the velocity of the bullet. To avoid regressing with the seam problem, check the dot product of the velocity with the collision normal.