Upgrades
Attributes: add ricochet, bulletCount. Currently, bulletCount is used to check which fire pattern to load in FlandrePC, and ricochet property will be used to initialize ricochetCount in PlayerBullet cons.
upgradeAttributes is now a map: it maps to UpgradeInfo struct, which desribes the amount of attribute increase per upgrade, and the sprite used.

Remove UpgradeImpl, instead use a single Upgrade class
Remove Upgrade::init(), conditional load is handled in the adapter

Player::setFirePatterns -> setFirePattern

PlayerInfo: display new upgrade attributes; use upgrade sprites as icons.

GState: remove per-player upgrades. Upgrades are each stored in a bitset<8>, and are stored as an array for all attributes.