getShieldCost() - Implemement shield level, add AttributeSystem to LuaAPI
AttributeSystem - get and set by name and ID
Attribute <--> string bimap

Shield level corresponds to the maximum angle that can be protected against, and also provides a discount for smaller angles that were already protected by the previous shield level.

shieldScalars - represents the discrete angle steps to use for the shield, corresponding to 45, 90, 135, and 180 angles.
shieldCosts - the cost to block a bullet within the angle threshold for that level, or -1.0 if it can't.

AreaSensor::updateBoss - isBossActive flag
Remove shieldCost
Give Flandre default shield level 1 and Cirno shield level 2