dodge
"sprint" action is now a dodge move, rather than a speed multiplier. Once activated, the player will move in the sprint direction (the direction of the left arrows when activated, otherwise in the player's current direction).
Activating sprint mode has a stamina cost.
Do not allow shield/focus when sprinting.
Sprint mode also needs to increase acceleration.
There is a recovery phase (equal to sprint time or half of cooldown time), where the player's speed is set to zero, but with the same sprint acceleration modifier - just long enough to bring her to a complete stop.