Remove GSpace scene/graphics single-threading "optimizations"
graphics/sprite/lightmap wrappers do not call the method in single thread mode, but use sceneActions regardless. These optimizations are unnecessary, as the spriteSpatialUpdate() already significantly reduces the number of sceneActions being run. Moreover, when run in single thread mode, it still runs the sceneActions afterwards, before updating the renders, in the same frame. There is a risk is that these optimizations will actually cause scene actions to run out of order in comparison with other code in the game logic that makes scene actions without using these wrappers.