MagicEffectSystem
There is now a single place that handles the lifecycle of MagicEffects, rather than storing this per-object.
refactor MagicEffect lists & updateMagicEffect out of GObject.
In other to ensure correctness, MagicEffectSytem::removeObjectEffects() needs to immediately unload (and destroy) magic effects associated with an object, before that object is deleted. Also, an object has the ability to queue the removal of other objects on its removal, hence we cannot simply removeObjectEffects() using all of the objects in GSpace::toRemove beforehand.