Refactor QueryShape - it now takes filter data.
The callback will only be called if the filter data is a match. This means removal of unnecessary checking of type & layers in my callbacks.
This also means the internal QueryShapeCallback will check filter collision before checking narrow-phase (i.e. actual fixture/shape collision).

refactor b2Filter to its own file, so it can be used in b2World methods
change its constructor so that it initialized categoryBits, maskBits to zero.
refactor code from b2ContactFilter::ShouldCollide -> b2Filter::shouldCollide()
isBlankType / isQueryCollide - in order to have filter data work for queries, we want it to either be able to represent the identity of an actual object (i.e. normal shouldCollide), or simply specifying a combination of object types & layers . If the query filter is a blank type (meaning categoryBits and groupIndex are zero), then it only checks type by comparing its maskBits against the categoryBits of the fixture being queried.