fix player sound issues
GObject/Player::playSoundSpatial override, so that the Player does not actually create spatial sounds. This returns to previous behavior, where the Player would not create spatial sound effects, as that would result in a sound that is zero distance from the sound source.
OpenAL-Soft supports creating zero-distance spatial sounds, except it produces artifacts if that source is moved when playing.