Refactor physics callbacks
The contact functions are now std::function's that simply take a b2Contact*.
This means two factory/adapter functions - one to handle existing object pairs, and one for radar sensors.
In both cases, the lambda remembers the order in which the parameters are declared, and account for fixture order.
AddHandler still sets the callback pair key with smaller numeric value first,  and BeginContact/EndContact still take adavantage of this, but this is no longer stored in the contact handler maps.