refactor physics handlers
makeObjectPairFunc is now templated, and will deduce the sub-types of the pair of GObjects. It will automatically check order, perform dynamic_cast, and ensure objects are valid.
This removes a lot of redundant dynamic casting and checking for null objects in the callback functions.
note: environmentAreaSensorEnd was likely never calling the appropriate end contact, since it was out of order.