Fix MagicEffectSystem memory error on removal of MagicEffect
The problem is that the deleted MagicEffect is not removed from updateEffects.
magicEffects.erase should be the last line in applyRemove(). Typically, when this is called, the shared_ptr for the MagicEffect decrements to zero and is immediately deallocated. Since magicEffects is now keyed by ID, this means (at least in debug) that the crnt (raw ptr) will immediately have garbage data, and the crnt ID will be invalid.