SpellSystem::logSpells
Log statements for spells' lifecycles.
-also fix format string in stopSpell warning log output.
refactor applyRemovals() from update(). It could be called in stopObjectSpells, since that is only supposed to be used in GSpace::processRemoval. However, according to naming convention, it merely stops the spells, rather than deleting them. Thus, it is simpler to call it in the destructor instead.

-Remove logging statement from IllusionDial script
-Remove second call to stopObjectSpells.