add hashset_gobject_ref to lua
It is simply a struct that contains an unordered_set<gobject_ref> and wraps methods. This avoids using the sol  wrappers for unordered_set.
The included sol wrapper for unordered_set sems to be having issues. " cannot perform a 'set': ... iterator reference is not writable ", but only when used for the ghost headstone sensor.
-IllusionDial was working, but has also been refactored

add sol::error catch in spaceUpdateMain

gobject_ref: remove operator== overload and add operator=