object_properties polymorphism
xml.hpp needs to load after GObject.hpp
shared_ptr.hpp downcast - perform dynamic_cast conversion
use a single objects table for all object properties

objectAdapter - refactor the various adapters to use the same name. also, add overload for properties types that aren't added as map construction adapters.
loadObjectsShared was strictly used for the different object properties types, and is refactored for this. It now uses the single objects table.
-It calls copyBaseObjectsShared after creating the result object, so now all objects have bsae object inheritence.
-It also calls the overload of objectAdapter - it checks if objectAdapter actually returns a value, for types that don't use an adapter.

getObjectProperties is basically identical to the getter functions, except they are more concise. A couple that aren't used have been removed.
It will get the props from the objects table and try to dynamic_cast it, giving warning messages if it is not in the table, or if the dynamic_cast fails.

parseObject(floorsegment_properties) now calls parseObject(object_properties) - currently, it doesn't make a difference, but eventually may be useful for type name and script cls name.