misc fixes
GSpace::pathToTile: the first point in the path should be the starting point.
fix addNavObstacle not being called in processAdditions. now the condition for whether an object is an obstacle is that it's in the ground layer and static.

make Patchouli2 kinematic and add pathToPoint to FollowPathKinematic.

Function: runEnter / runUpdate / runExit interface. Checks if state is valid and sets it as necessary.
FollowPathKinematic: init is supposed to be onEnter
-also, add loop parameter