move call to initializeBody from GObject constructor to GSpace::processAdditions
Now it should be safe to call createObject anywhere, including inside of a physics callback.
In the event it calls a getter for a physics attribute before the body is loaded, return the starting attribute instead.
Add some checks in case in case it tries to set a physics attribute when the body hasn't been loaded yet.